new gameboard
this code creates an array that stores all information about the state
of the board, including whether each square is available, occupied, or
unavailable.  the script draws each square according to the data stored
in the array.

i deleted the extraneous content from the html file for now.  the
canvas should keep us busy enough for a bit!